server/requirements.txt
Django>=3.2,<4.0
djongo>=1.3.6
pymongo>=3.12.3,<4.0
django-cors-headers>=3.10.0
djangorestframework>=3.12.4
djangorestframework-simplejwt>=5.2.0
PyJWT>=2.3.0
gunicorn>=20.1.0
python-dotenv>=0.19.0
requests>=2.26.0
llama-cpp-python>=0.2.0
pydantic>=2.0.0
groq==0.4.1
